requirements/doc_requirements.txt
# Documentation
sphinx
numpydoc
myst-nb
sphinx-design
sphinx-book-theme
sphinxcontrib-bibtex
sphinxcontrib-mermaid
sphinx-automodapi
graphviz
# Plotting
matplotlib
